feat: add nullable attributes and action on merge request
In this commit, the types of labels, assignees, and assignee have been changed to be nullable in the MergeRequest and Issue classes. This change has been made to better handle cases where this data is not present in the push event, making it more flexible for different real-life scenarios. An action type was also added to the MergeRequest to track the kind of performed action.